Add code until end of Chapter 4

+++ src/process.cpp
@@ -4,25 +4,52 @@
 #include <unistd.h>
 
 #include <libsdb/error.hpp>
+#include <libsdb/pipe.hpp>
 #include <libsdb/process.hpp>
 
-std::unique_ptr<sdb::process> sdb::process::launch(std::filesystem::path path) {
+namespace {
+void exit_with_perror(sdb::pipe& channel, std::string const& prefix) {
+  auto message = prefix + ": " + std::strerror(errno);
+  channel.write(reinterpret_cast<std::byte*>(message.data()), message.size());
+  exit(EXIT_FAILURE);
+}
+}  // namespace
+
+std::unique_ptr<sdb::process> sdb::process::launch(std::filesystem::path path,
+                                                   bool debug) {
+  pipe channel{/*close_on_exec=*/true};
   pid_t pid;
   if ((pid = fork()) < 0) {
     error::send_errno("fork failed");
   }
 
   if (pid == 0) {
-    if (ptrace(PTRACE_TRACEME, 0, nullptr, nullptr) < 0) {
-      error::send_errno("Tracing failed");
+    channel.close_read();
+    sleep(2);
+    if (debug && ptrace(PTRACE_TRACEME, 0, nullptr, nullptr) < 0) {
+      exit_with_perror(channel, "Tracing failed");
     }
     if (execlp(path.c_str(), path.c_str(), nullptr) < 0) {
-      error::send_errno("exec failed");
+      exit_with_perror(channel, "exec failed");
     }
   }
 
-  std::unique_ptr<process> proc(new process(pid, /*terminate_on_end=*/true));
-  proc->wait_on_signal();
+  channel.close_write();
+  auto data = channel.read();
+  channel.close_read();
+
+  if (data.size() > 0) {
+    waitpid(pid, nullptr, 0);
+    auto chars = reinterpret_cast<char*>(data.data());
+    error::send(std::string(chars, chars + data.size()));
+  }
+
+  std::unique_ptr<process> proc(
+      new process(pid, /*terminate_on_end=*/true, debug));
+  if (debug) {
+    proc->wait_on_signal();
+  }
+
   return proc;
 }
 
@@ -34,7 +61,8 @@ std::unique_ptr<sdb::process> sdb::process::attach(pid_t pid) {
     error::send_errno("Could not attach");
   }
 
-  std::unique_ptr<process> proc(new process(pid, /*terminate_on_end=*/false));
+  std::unique_ptr<process> proc(
+      new process(pid, /*terminate_on_end=*/false, /*attached=*/true));
   proc->wait_on_signal();
   return proc;
 }
@@ -42,13 +70,14 @@ std::unique_ptr<sdb::process> sdb::process::attach(pid_t pid) {
 sdb::process::~process() {
   if (pid_ != 0) {
     int status;
-    if (state_ == sdb::process_state::running) {
-      kill(pid_, SIGSTOP);
-      waitpid(pid_, &status, 0);
+    if (is_attached_) {
+      if (state_ == sdb::process_state::running) {
+        kill(pid_, SIGSTOP);
+        waitpid(pid_, &status, 0);
+      }
+      ptrace(PTRACE_DETACH, pid_, nullptr, nullptr);
+      kill(pid_, SIGCONT);
     }
-    ptrace(PTRACE_DETACH, pid_, nullptr, nullptr);
-    kill(pid_, SIGCONT);
-
     if (terminate_on_end_) {
       kill(pid_, SIGKILL);
       waitpid(pid_, &status, 0);
